+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove autofocus on measurement value field
 - Unify segmented time span controls on dashboard and habit pages
 - Increase analyzed habit completion time span to 90 days
+- Show date in habit completion rate chart tooltip
 
 ## [0.8.205] - 2022-12-09
 ### Added: